:wrench: Config to build
