+++ tsconfig.cj.json
@@ -1,0 +1,10 @@
+{
+    "extends": "./tsconfig.json",
+    "exclude": [
+        "src/dev"
+    ],
+    "compilerOptions": {
+        "module": "CommonJS",
+        "outDir": "dist/cj"
+    }
+}
